requirements.txt
langchain_experimental==0.3.4
chromadb
jq
langchain
langchain-text-splitters
langchain_community
langchain-openai
langchain-core
langchain-openai
langgraph
pypdf
tiktoken
scikit-learn
streamlit
mcp
pydantic
python-dotenv
openai
pyats[full]
sentence-transformers
foundry-local-sdk